requirements.txt
azure-iot-device==2.12.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
deprecation==2.1.0
Flask==3.0.0
gevent==23.9.1
greenlet==3.0.2
idna==3.6
itsdangerous==2.1.2
janus==1.0.0
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
paho-mqtt==1.6.1
PySocks==1.7.1
requests==2.31.0
requests-unixsocket==0.3.0
typing_extensions==4.9.0
urllib3==1.26.18
websocket==0.2.1
websockets==12.0
Werkzeug==3.0.1
zope.event==5.0
zope.interface==6.1
